pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "google-analytics-mcp"
version = "2.0.0" # Major version bump for refactoring
description = "Google Analytics 4 MCP Server with Smart Optimizations - Access GA4 data in Claude, Cursor and other MCP clients. Prevents context window crashes with intelligent data volume management."
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Surendran B", email = "reachsuren@gmail.com"}
]
requires-python = ">=3.10"
dependencies = [
    "mcp>=1.2.0",
    "google-analytics-data==0.19.0",
    "google-auth==2.40.0"
]
keywords = ["google-analytics", "mcp", "ai-assistant", "analytics", "ga4", "claude", "cursor", "windsurf"]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.urls]
Homepage = "https://github.com/surendranb/google-analytics-mcp"
Repository = "https://github.com/surendranb/google-analytics-mcp"
Issues = "https://github.com/surendranb/google-analytics-mcp/issues"

[project.scripts]
ga4-mcp-server = "ga4_mcp.server:main"

[tool.setuptools]
packages = ["ga4_mcp", "ga4_mcp.tools"]
